Refactor GitHub API models into new module
This now distinguishes between generic resources available from the
GitHub REST API and webhook payloads (which may contain REST resources).
